LoadConfig bug with float properties parsing.
Former-commit-id: b22f05e8cfc42510303d2fa49b7512030131ce0b
Former-commit-id: 2a76338d51b2ebc24a94a42f2a5bc7f2043d8c01
Former-commit-id: 968b19277726cdb8fac94d0be5fd3dda4fc41467

+++ vizia_api_src/ViziaDoomGame.cpp
@@ -1048,12 +1048,12 @@ namespace Vizia {
             try{
                 if (key =="living_reward" or key =="livingreward"){
                     float value = boost::lexical_cast<float>(val);
-                    this->setLivingReward((unsigned int)value);
+                    this->setLivingReward(value);
                     continue;
                 }
                 if (key == "deathpenalty" or key == "death_penalty"){
                     float value = boost::lexical_cast<float>(val);
-                    this->setDeathPenalty((unsigned int)value);
+                    this->setDeathPenalty(value);
                     continue;
                 }
